requirements.txt
Django>=4.2
django-jet-reboot
python-dotenv
requests
psycopg2
